pyproject.toml
[tool.poetry]
name = "dynamicprogramwithchatgpt"
version = "0.1.0"
description = "ChatGPT with coding powers"
authors = ["jorgegomez <jorgegomez@bmat.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.6"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
